third_party/WebKit/LayoutTests/http/tests/w3c/webperf/approved/UserTiming/idlharness-expected.txt

idlharness test

This test validates the WebIDL included in the user Timing specification.

partial interface Performance {
    void mark(DOMString markName);
    void clearMarks(optional  DOMString markName);

    void measure(DOMString measureName, optional DOMString startMark, optional DOMString endMark);
    void clearMeasures(optional DOMString measureName);
};

interface PerformanceMark : PerformanceEntry {
};

interface PerformanceMeasure : PerformanceEntry {
};


FAIL Performance interface: operation mark(DOMString) assert_true: property is not configurable expected true got false
FAIL Performance interface: operation clearMarks(DOMString) assert_true: property is not configurable expected true got false
FAIL Performance interface: operation measure(DOMString,DOMString,DOMString) assert_true: property is not configurable expected true got false
FAIL Performance interface: operation clearMeasures(DOMString) assert_true: property is not configurable expected true got false
PASS Performance must be primary interface of window.performance 
PASS Stringification of window.performance 
PASS Performance interface: window.performance must inherit property "mark" with the proper type (0) 
PASS Performance interface: calling mark(DOMString) on window.performance with too few arguments must throw TypeError 
PASS Performance interface: window.performance must inherit property "clearMarks" with the proper type (1) 
PASS Performance interface: calling clearMarks(DOMString) on window.performance with too few arguments must throw TypeError 
PASS Performance interface: window.performance must inherit property "measure" with the proper type (2) 
PASS Performance interface: calling measure(DOMString,DOMString,DOMString) on window.performance with too few arguments must throw TypeError 
PASS Performance interface: window.performance must inherit property "clearMeasures" with the proper type (3) 
PASS Performance interface: calling clearMeasures(DOMString) on window.performance with too few arguments must throw TypeError 
FAIL PerformanceMark interface: existence and properties of interface object assert_equals: prototype of window's property "PerformanceMark" is not Function.prototype expected function "function Empty() {}" but got function "function PerformanceEntry() { [native code] }"
PASS PerformanceMark interface object length 
FAIL PerformanceMark interface: existence and properties of interface prototype object assert_equals: class string of PerformanceMark.prototype expected "[object PerformanceMarkPrototype]" but got "[object Object]"
PASS PerformanceMark interface: existence and properties of interface prototype object's "constructor" property 
FAIL PerformanceMeasure interface: existence and properties of interface object assert_equals: prototype of window's property "PerformanceMeasure" is not Function.prototype expected function "function Empty() {}" but got function "function PerformanceEntry() { [native code] }"
PASS PerformanceMeasure interface object length 
FAIL PerformanceMeasure interface: existence and properties of interface prototype object assert_equals: class string of PerformanceMeasure.prototype expected "[object PerformanceMeasurePrototype]" but got "[object Object]"
PASS PerformanceMeasure interface: existence and properties of interface prototype object's "constructor" property 
